Revert "Feat/week 008 api clean architecture with getx 2 (#9)"
This reverts commit 91eae13e3dd13960c969bd196d9fd1c5a49f0a68.
