--- lib/utils/format.dart
@@ -1,10 +1,0 @@
-import 'package:intl/intl.dart';
-
-String formatRupiah(num value) {
-  final f = NumberFormat.currency(
-    locale: 'id_ID',
-    symbol: 'Rp ',
-    decimalDigits: 0,
-  );
-  return f.format(value);
-}
