--- lib/week_8/features/courses/domain/repositories/course_repository.dart
@@ -1,18 +1,0 @@
-import '../entities/course.dart';
-
-abstract class CourseRepository {
-  Future<List<Course>> getCourses();
-  Future<Course> getCourseById(String id); // NEW
-  Future<Course> createCourse({
-    required String title,
-    required int price,
-    required String category,
-  });
-  Future<Course> updateCourse({
-    required String id,
-    required String title,
-    required int price,
-    required String category,
-  });
-  Future<void> deleteCourse(String id);
-}
