--- lib/week_8/features/courses/domain/use_cases/delete_course_use_case.dart
@@ -1,7 +1,0 @@
-import '../repositories/course_repository.dart';
-
-class DeleteCourseUseCase {
-  final CourseRepository repo;
-  DeleteCourseUseCase(this.repo);
-  Future<void> call(String id) => repo.deleteCourse(id);
-}
