--- lib/week_8/features/courses/domain/use_cases/get_course_by_id_use_case.dart
@@ -1,9 +1,0 @@
-import '../entities/course.dart';
-import '../repositories/course_repository.dart';
-
-class GetCourseByIdUseCase {
-  final CourseRepository repo;
-  GetCourseByIdUseCase(this.repo);
-
-  Future<Course> call(String id) => repo.getCourseById(id);
-}
