--- lib/week_8/features/courses/domain/use_cases/get_courses_use_case.dart
@@ -1,8 +1,0 @@
-import '../entities/course.dart';
-import '../repositories/course_repository.dart';
-
-class GetCoursesUseCase {
-  final CourseRepository repo;
-  GetCoursesUseCase(this.repo);
-  Future<List<Course>> call() => repo.getCourses();
-}
