--- lib/week_8/features/courses/presentation/pages/new_kelas_result.dart
@@ -1,12 +1,0 @@
-class NewKelasResult {
-  final String nama;
-  final int harga;
-  final String kategori;
-  final String? thumbnail;
-  const NewKelasResult({
-    required this.nama,
-    required this.harga,
-    required this.kategori,
-    this.thumbnail,
-  });
-}
